+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/FrameDescriptorTest.java
@@ -104,11 +104,9 @@ public class FrameDescriptorTest {
         assertEquals(2, d.getSize());
         assertEquals("i2", d.getSlots().get(1).getInfo());
         assertEquals(FrameSlotKind.Float, d.getFrameSlotKind(d.getSlots().get(1)));
-        assertEquals(1, d.getSlots().get(1).getIndex());
 
         FrameDescriptor copy = d.copy();
         assertEquals(2, copy.getSize());
-        assertEquals(1, copy.getSlots().get(1).getIndex());
         assertEquals("Info is copied", "i2", copy.getSlots().get(1).getInfo());
         assertEquals("Kind isn't copied", FrameSlotKind.Illegal, copy.getFrameSlotKind(copy.getSlots().get(1)));
     }